+++ doc/version_history.rst
@@ -6,6 +6,15 @@
 Version History
 ##################
 
+.. _lsst.ts.hexgui-0.2.0:
+
+-------------
+0.2.0
+-------------
+
+* Read the configuration file in **ts_config_mttcs**.
+* Add the **signals.py**, **config.py**, and **status.py**.
+
 .. _lsst.ts.hexgui-0.1.0:
 
 -------------